(refactor) finalize PR, bump to 4.0.15
