+++ .eslintrc.json
@@ -1,6 +1,9 @@
 {
+  "root": true,
+
   "plugins": [
-    "jsdoc"
+    "jsdoc",
+    "mocha"
   ],
 
   "env": {
@@ -17,6 +20,7 @@
     "plugin:import/errors",
     "plugin:import/warnings",
     "plugin:lodash/recommended",
+    "plugin:mocha/recommended",
     "plugin:jsdoc/recommended"
   ],
 
@@ -25,11 +29,16 @@
   },
 
   "rules": {
+    "jsdoc/check-tag-names": 0,
+    "jsdoc/no-undefined-types": 0,
     "no-multiple-empty-lines": ["error", { "max": 1, "maxEOF": 1 }],
     "promise/no-callback-in-promise": 0,
     "lodash/prefer-lodash-method": 0,
     "lodash/prefer-noop": 0,
     "lodash/prefer-constant": 0,
-    "node/no-unpublished-require": 0
+    "node/no-unpublished-require": 0,
+    "mocha/no-mocha-arrows": 0,
+    "mocha/no-setup-in-describe": 0,
+    "mocha/no-skipped-tests": 0
   }
 }
