--- .jsdoc.json
@@ -1,61 +1,0 @@
-{
-  "tags": {
-    "allowUnknownTags": false,
-    "dictionaries": ["jsdoc"]
-  },
-
-  "plugins": [
-    "plugins/markdown",
-    "jsdoc-mermaid",
-    "jsdoc-plugin-npm",
-    "jsdoc-plugin-url"
-  ],
-
-  "source": {
-    "include": ["index.js", "lib", "LICENSE.md", "README.md"],
-    "includePattern": ".js$"
-  },
-
-  "opts": {
-    "destination": "./docs/jsdoc",
-    "encoding": "utf8",
-    "recurse": true,
-    "verbose": true,
-    "template": "./node_modules/docdash"
-  },
-
-  "markdown": {
-    "parser": "gfm",
-    "hardwrap": false,
-    "idInHeadings": true
-  },
-
-  "templates": {
-    "cleverLinks": true,
-    "default": {
-      "outputSourceFiles": true,
-      "includeDate": true,
-      "useLongnameInNav": false
-    }
-  },
-
-  "docdash": {
-    "static": true,
-    "sort": true,
-    "search": true,
-    "collapse": true,
-    "wrap": true,
-    "typedefs": true,
-    "private": false,
-    "sectionOrder": [
-      "Tutorials",
-      "Namespaces",
-      "Modules",
-      "Classes",
-      "Externals",
-      "Events",
-      "Mixins",
-      "Interfaces"
-    ]
-  }
-}
