+++ CHANGELOG
@@ -1,10 +1,7 @@
-4.0.16
-- meta: update JSDoc doclets for bfx-oss-docs
-
 4.0.15
-- meta: bump all dependencies
-- meta: update babel config
-- refactor: scoped all JSDoc data under module:bitfinex-api-node
+- meta: bump deps
+- meta: normalize eslint config
+- refactor: improve doclets
 
 4.0.14
 - fix: README docs reference
