+++ lib/client_manager.js
@@ -1,18 +1,14 @@
 'use strict'
 
+const { RESTv1, RESTv2 } = require('bfx-api-node-rest')
+const WSv1 = require('bfx-api-node-ws1')
 const WSv2 = require('./ws2')
 
 /**
  * Provides access to versions 1 & 2 of the HTTP & WebSocket Bitfinex APIs.
- * Main export of {@npm bitfinex-api-node}.
  *
- * @see module:bitfinex-api-node.WSv2
- * @see module:bfx-api-node-ws1.WSv1
- * @see module:bfx-api-node-rest.RESTv2
- * @see module:bfx-api-node-rest.RESTv1
- *
+ * @see {@link WSv2}
  * @class
- * @memberof module:bitfinex-api-node
  */
 class ClientManager {
   /**
@@ -22,8 +18,7 @@ class ClientManager {
    * @param {string} opts.authToken - optional auth option
    * @param {string} opts.transform - if true, packets are converted to models
    * @param {string} opts.nonceGenerator - optional
-   * @param {module:bitfinex-api-node.WSv2~Configuration} opts.ws - ws
-   *   transport options
+   * @param {WSv2~Configuration} opts.ws - ws transport options
    * @param {string} opts.rest - rest transport options
    */
   constructor (opts = {
@@ -79,8 +74,8 @@ class ClientManager {
    * @param {number} [version=2] - 1 or 2
    * @param {object} [extraOpts={}] - passed to transport constructor
    * @returns {(
-   *   module:bfx-api-node-rest.RESTv1|
-   *   module:bfx-api-node-rest.RESTv2
+   *   bfx-api-node-rest.RESTv1|
+   *   bfx-api-node-rest.RESTv2
    * )} transport
    */
   rest (version = 2, extraOpts = {}) {
@@ -106,12 +101,9 @@ class ClientManager {
    * Returns a new WebSocket API class instance (cached by version)
    *
    * @param {number} [version=2] - 1 or 2
-   * @param {module:bitfinex-api-node.WSv2~Configuration} [extraOpts={}] -
-   *   passed to transport constructor
-   * @returns {(
-   *   module:bitfinex-api-node.WSv2|
-   *   module:bfx-api-node-ws1.WSv1
-   * )} transport
+   * @param {WSv2~Configuration} [extraOpts={}] - passed to transport
+   *   constructor
+   * @returns {WSv2|bfx-api-node-ws1.WSv1} transport
    */
   ws (version = 2, extraOpts = {}) {
     if (version !== 1 && version !== 2) {
