+++ lib/types/candle_channel.js
@@ -1,10 +1,9 @@
 'use strict'
 
 /**
- * An object describing a trade data channel on a
- * {@link module:bitfinex-api-node.WSv2|WSv2} data stream.
+ * An object describing a trade data channel on a {@link WSv2} data stream.
  *
- * @typedef {object} module:bitfinex-api-node.TradeChannel
+ * @typedef {object} CandleChannel
  * @property {string} channel - 'candles'
  * @property {number} chanId - channel ID
  * @property {string} key - data key, 'trade:TF:SYMBOL' i.e. 'trade:1m:tBTCUSD'
