+++ lib/types/order_book_channel.js
@@ -1,10 +1,9 @@
 'use strict'
 
 /**
- * An object describing an order book data channel on a
- * {@link module:bitfinex-api-node.WSv2|WSv2} data stream.
+ * An object describing an order book data channel on a WSv2 data stream.
  *
- * @typedef {object} module:bitfinex-api-node.OrderBookChannel
+ * @typedef {object} OrderBookChannel
  * @property {string} channel - 'book'
  * @property {number} chanId - channel ID
  * @property {string} symbol - symbol
