+++ lib/types/ticker_channel.js
@@ -1,10 +1,9 @@
 'use strict'
 
 /**
- * An object describing a ticker data channel on a
- * {@link module:bitfinex-api-node.WSv2|WSv2} data stream.
+ * An object describing a ticker data channel on a WSv2 data stream.
  *
- * @typedef {object} module:bitfinex-api-node.TickerChannel
+ * @typedef {object} TickerChannel
  * @property {string} channel - 'ticker'
  * @property {number} chanId - channel ID
  * @property {string} symbol - symbol
