+++ lib/types/ws2_configuration.js
@@ -1,10 +1,9 @@
 'use strict'
 
 /**
- * Configuration options for a {@link module:bitfinex-api-node.WSv2|WSv2}
- * client instance.
+ * Configuration options for a WSv2 client instance.
  *
- * @typedef {object} module:bitfinex-api-node.WSv2~Configuration
+ * @typedef {object} WSv2~Configuration
  * @property {string} [affCode=''] - affiliate code to be applied to all orders
  * @property {string} [apiKey=''] - API key
  * @property {string} [apiSecret=''] - API secret
