--- lib/util/index.js
@@ -1,9 +1,0 @@
-'use strict'
-
-const isClass = require('./is_class')
-const isSnapshot = require('./is_snapshot')
-
-module.exports = {
-  isClass,
-  isSnapshot
-}
