--- lib/util/is_class.js
@@ -1,12 +1,0 @@
-'use strict'
-
-const _isFunction = require('lodash/isFunction')
-
-const isClass = (f) => {
-  return (
-    (_isFunction(f)) &&
-    (/^class\s/.test(Function.prototype.toString.call(f)))
-  )
-}
-
-module.exports = isClass
