--- lib/util/is_snapshot.js
@@ -1,5 +1,0 @@
-'use strict'
-
-const isSnapshot = msg => msg[0] && Array.isArray(msg[0])
-
-module.exports = isSnapshot
