+++ lib/ws2.js
@@ -5,18 +5,18 @@ const debug = require('debug')('bfx:ws2')
 const WebSocket = require('ws')
 const Promise = require('bluebird')
 const CbQ = require('cbq')
-const _Throttle = require('lodash.throttle') // eslint-disable-line
+const _pick = require('lodash/pick')
+const _isEqual = require('lodash/isEqual')
 const _isArray = require('lodash/isArray')
 const _isEmpty = require('lodash/isEmpty')
 const _isString = require('lodash/isString')
 const _isNumber = require('lodash/isNumber')
 const _includes = require('lodash/includes')
-const _pick = require('lodash/pick')
-const _isEqual = require('lodash/isEqual')
 const _isFinite = require('lodash/isFinite')
+const _findLast = require('lodash/findLast')
+const _Throttle = require('lodash.throttle') // eslint-disable-line
 const { genAuthSig, nonce } = require('bfx-api-node-util')
 const LosslessJSON = require('lossless-json')
-const getMessagePayload = require('./util/ws2')
 
 const {
   BalanceInfo,
@@ -38,18 +38,27 @@ const {
   FundingTicker
 } = require('bfx-api-node-models')
 
-require('./types/promise_throttle')
-
 const DATA_CHANNEL_TYPES = ['ticker', 'book', 'candles', 'trades']
 const UCM_NOTIFICATION_TYPE = 'ucm-notify-ui'
 const MAX_CALC_OPS = 8
 
 /**
+ * Resolves the message payload; useful for getting around sequence numbers
+ *
+ * @private
+ *
+ * @param {Array} msg - message to parse
+ * @returns {Array} payload - undefined if not found
+ */
+const getMessagePayload = (msg = []) => {
+  return _findLast(msg, i => _isArray(i))
+}
+
+/**
  * Communicates with v2 of the Bitfinex WebSocket API
  *
  * @class
- * @memberof module:bitfinex-api-node
- * @augments EventEmitter
+ * @augments events.EventEmitter
  */
 class WSv2 extends EventEmitter {
   /**
@@ -68,7 +77,7 @@ class WSv2 extends EventEmitter {
    *
    * await ws.open()
    *
-   * @param {module:bitfinex-api-node.WSv2~Configuration} [config] - client
+   * @param {WSv2~Configuration} [config] - client
    *   configuration
    */
   constructor (config = {
@@ -156,12 +165,12 @@ class WSv2 extends EventEmitter {
 
   /**
    * Set `calc` and `dms` values to be used on the next
-   * {@link module:bitfinex-api-node.WSv2#auth} call
+   * {@link WSv2#auth} call
    *
    * @param {object} args - arguments
    * @param {number} [args.calc] - calc value
    * @param {number} [args.dms] - dms value, active 4
-   * @see module:bitfinex-api-node.WSv2#auth
+   * @see {@link WSv2#auth}
    */
   updateAuthArgs (args = {}) {
     this._authArgs = {
@@ -174,8 +183,8 @@ class WSv2 extends EventEmitter {
    * Fetch the current default auth parameters
    *
    * @returns {object} authArgs
-   * @see module:bitfinex-api-node.WSv2#updateAuthArgs
-   * @see module:bitfinex-api-node.WSv2#auth
+   * @see {@link WSv2#updateAuthArgs}
+   * @see {@link WSv2#auth}
    */
   getAuthArgs () {
     return this._authArgs
@@ -183,11 +192,11 @@ class WSv2 extends EventEmitter {
 
   /**
    * Update the internal API credentials, used on subsequent
-   * {@link module:bitfinex-api-node.WSv2#auth} calls
+   * {@link WSv2#auth} calls
    *
    * @param {string} apiKey - API key
    * @param {string} apiSecret - API secret
-   * @see module:bitfinex-api-node.WSv2#auth
+   * @see {@link WSv2#auth}
    */
   setAPICredentials (apiKey, apiSecret) {
     this._apiKey = apiKey
@@ -199,10 +208,10 @@ class WSv2 extends EventEmitter {
    * subscribed too.
    *
    * @returns {number} count
-   * @see module:bitfinex-api-node.WSv2#subscribeTrades
-   * @see module:bitfinex-api-node.WSv2#subscribeTicker
-   * @see module:bitfinex-api-node.WSv2#subscribeCandles
-   * @see module:bitfinex-api-node.WSv2#subscribeOrderBook
+   * @see {@link WSv2#subscribeTrades}
+   * @see {@link WSv2#subscribeTicker}
+   * @see {@link WSv2#subscribeCandles}
+   * @see {@link WSv2#subscribeOrderBook}
    */
   getDataChannelCount () {
     return Object
@@ -227,7 +236,7 @@ class WSv2 extends EventEmitter {
    * @param {string} channel - channel type
    * @param {string} identifier - unique identifier for the reference
    * @returns {boolean} hasRef
-   * @see module:bitfinex-api-node.WSv2#managedSubscribe
+   * @see {@link WSv2#managedSubscribe}
    */
   hasSubscriptionRef (channel, identifier) {
     const key = `${channel}:${identifier}`
@@ -402,10 +411,13 @@ class WSv2 extends EventEmitter {
 
   /**
    * @private
+   *
+   * @returns {Promise} p
    */
   async reconnectAfterClose () {
     if (!this._isReconnecting || this._ws !== null || this._isOpen) {
-      return this.reconnect()
+      this.reconnect()
+      return
     }
 
     await this.open()
@@ -920,7 +932,7 @@ class WSv2 extends EventEmitter {
    * await ws.open()
    *
    * @param {string} symbol - symbol for order book
-   * @returns {module:bfx-api-node-models.OrderBook} ob - null if not found
+   * @returns {bfx-api-node-models.OrderBook} ob - null if not found
    */
   getOB (symbol) {
     if (!this._orderBooks[symbol]) return null
@@ -936,7 +948,7 @@ class WSv2 extends EventEmitter {
    * Set `manageOrderBooks: true` in the constructor to use.
    *
    * @param {string} symbol - symbol for order book
-   * @returns {module:bfx-api-node-models.OrderBook} ob - null if not found
+   * @returns {bfx-api-node-models.OrderBook} ob - null if not found
    */
   getLosslessOB (symbol) {
     if (!this._losslessOrderBooks[symbol]) return null
@@ -1119,8 +1131,8 @@ class WSv2 extends EventEmitter {
    *
    * @param {string} key - key for candle set
    * @returns {(
-   *   module:bfx-api-node-models.Candle|
-   *   module:bfx-api-node-models.Candle~ArrayData[]
+   *   bfx-api-node-models.Candle|
+   *   bfx-api-node-models.Candle~ArrayData[]
    * )} candles - empty array if none exist
    */
   getCandles (key) {
@@ -1423,10 +1435,10 @@ class WSv2 extends EventEmitter {
    * @todo will be refactored to return promise from subscribe() call instead
    *   of sub action taken flag
    *
-   * @see module:bitfinex-api-node.WSv2#subscribeTrades
-   * @see module:bitfinex-api-node.WSv2#subscribeTicker
-   * @see module:bitfinex-api-node.WSv2#subscribeCandles
-   * @see module:bitfinex-api-node.WSv2#subscribeOrderBook
+   * @see {@link WSv2#subscribeTrades}
+   * @see {@link WSv2#subscribeTicker}
+   * @see {@link WSv2#subscribeCandles}
+   * @see {@link WSv2#subscribeOrderBook}
    * @example
    * const ws = new WSv2()
    *
@@ -1567,7 +1579,7 @@ class WSv2 extends EventEmitter {
    * connection. When enabled, the seq number will be the last value of
    * channel packet arrays.
    *
-   * @see module:bitfinex-api-node.WSv2#flags
+   * @see {@link WSv2#flags}
    *
    * @param {object} args - params
    * @param {boolean} [args.audit] - if true, an error is emitted on invalid seq
@@ -1582,7 +1594,7 @@ class WSv2 extends EventEmitter {
   /**
    * Enables a configuration flag.
    *
-   * @see module:bitfinex-api-node.WSv2#flags
+   * @see {@link WSv2#flags}
    * @example
    * const ws = new WSv2()
    *
@@ -1622,7 +1634,7 @@ class WSv2 extends EventEmitter {
   /**
    * Checks local state, relies on successful server config responses
    *
-   * @see module:bitfinex-api-node.WSv2#enableFlag
+   * @see {@link WSv2#enableFlag}
    *
    * @param {number} flag - flag to check for
    * @returns {boolean} enabled
@@ -1698,7 +1710,7 @@ class WSv2 extends EventEmitter {
   /**
    * Subscribe to a ticker data channel
    *
-   * @see module:bitfinex-api-node.WSv2#managedSubscribe
+   * @see {@link WSv2#managedSubscribe}
    * @example
    * await ws.subscribeTicker('tBTCUSD')
    *
@@ -1712,7 +1724,7 @@ class WSv2 extends EventEmitter {
   /**
    * Subscribe to a trades data channel
    *
-   * @see module:bitfinex-api-node.WSv2#managedSubscribe
+   * @see {@link WSv2#managedSubscribe}
    * @example
    * await ws.subscribeTrades('tBTCUSD')
    *
@@ -1726,7 +1738,7 @@ class WSv2 extends EventEmitter {
   /**
    * Subscribe to an order book data channel
    *
-   * @see module:bitfinex-api-node.WSv2#managedSubscribe
+   * @see {@link WSv2#managedSubscribe}
    * @example
    * await ws.subscribeOrderBook('tBTCUSD', 'R0', '25')
    *
@@ -1742,7 +1754,7 @@ class WSv2 extends EventEmitter {
   /**
    * Subscribe to a candle data channel
    *
-   * @see module:bitfinex-api-node.WSv2#managedSubscribe
+   * @see {@link WSv2#managedSubscribe}
    * @example
    * await ws.subscribeCandles('trade:5m:tBTCUSD')
    *
@@ -1756,7 +1768,7 @@ class WSv2 extends EventEmitter {
   /**
    * Subscribe to a status data channel
    *
-   * @see module:bitfinex-api-node.WSv2#managedSubscribe
+   * @see {@link WSv2#managedSubscribe}
    * @example
    * await ws.subscribeStatus('liq:global')
    *
@@ -1789,7 +1801,7 @@ class WSv2 extends EventEmitter {
   /**
    * Unsubscribe from a ticker data channel
    *
-   * @see module:bitfinex-api-node.WSv2#subscribeTicker
+   * @see {@link WSv2#subscribeTicker}
    * @example
    * await ws.unsubscribeTicker('tBTCUSD')
    *
@@ -1803,7 +1815,7 @@ class WSv2 extends EventEmitter {
   /**
    * Unsubscribe from a trades data channel
    *
-   * @see module:bitfinex-api-node.WSv2#subscribeTrades
+   * @see {@link WSv2#subscribeTrades}
    * @example
    * await ws.unsubcribeTrades('tBTCUSD')
    *
@@ -1817,7 +1829,7 @@ class WSv2 extends EventEmitter {
   /**
    * Unsubscribe from an order book data channel
    *
-   * @see module:bitfinex-api-node.WSv2#subscribeOrderBook
+   * @see {@link WSv2#subscribeOrderBook}
    * @example
    * await ws.unsubcribeOrderBook('tBTCUSD')
    *
@@ -1829,7 +1841,7 @@ class WSv2 extends EventEmitter {
   }
 
   /**
-   * @see module:bitfinex-api-node.WSv2#subscribeCandles
+   * @see {@link WSv2#subscribeCandles}
    * @example
    *   await ws.unsubscribeCandles('tBTCUSD', '1m')
    *
@@ -1842,7 +1854,7 @@ class WSv2 extends EventEmitter {
   }
 
   /**
-   * @see module:bitfinex-api-node.WSv2#subscribeStatus
+   * @see {@link WSv2#subscribeStatus}
    *
    * @param {string} key - key that was used in initial {@link WSv2#subscribeStatus} call
    * @returns {boolean} unsubscribed
@@ -1895,9 +1907,9 @@ class WSv2 extends EventEmitter {
    * order submit is confirmed. Emits an error if not authenticated. The order
    * can be either an array, key/value map, or Order object instance.
    *
-   * @see module:bitfinex-api-node.WSv2#cancelOrders
-   * @see module:bitfinex-api-node.WSv2#cancelOrder
-   * @see module:bitfinex-api-node.WSv2#updateOrder
+   * @see {@link WSv2#cancelOrders}
+   * @see {@link WSv2#cancelOrder}
+   * @see {@link WSv2#updateOrder}
    *
    * @example
    * const o = new Order({
@@ -1913,8 +1925,8 @@ class WSv2 extends EventEmitter {
    * console.log('order confirmed! status: %s', o.status)
    *
    * @param {(
-   *   module:bfx-api-node-models.Order|
-   *   module:bfx-api-node-models.Order~Data
+   *   bfx-api-node-models.Order|
+   *   bfx-api-node-models.Order~Data
    * )} order - order
    *
    * @returns {Promise} p - resolves on submit notification
@@ -1948,7 +1960,7 @@ class WSv2 extends EventEmitter {
    * the price queue. The changeset must contain the order ID, and supports a
    * 'delta' key to increase/decrease the total amount.
    *
-   * @see module:bitfinex-api-node.WSv2#submitOrder
+   * @see {@link WSv2#submitOrder}
    *
    * @param {object} changes - requires at least an 'id'
    * @returns {Promise} p - resolves on receiving an confirmation notification
@@ -1972,11 +1984,11 @@ class WSv2 extends EventEmitter {
    * confirmed. Emits an error if not authenticated. The ID can be passed as a
    * number, or taken from an order array/object.
    *
-   * @see module:bitfinex-api-node.WSv2#submitOrder
+   * @see {@link WSv2#submitOrder}
    *
    * @param {(
-   *   module:bfx-api-node-models.Order|
-   *   module:bfx-api-node-models.Order~Data|
+   *   bfx-api-node-models.Order|
+   *   bfx-api-node-models.Order~Data|
    *   number
    * )} order - order model, array, or ID to be cancelled
    *
@@ -2003,11 +2015,11 @@ class WSv2 extends EventEmitter {
    * Cancels multiple orders, returns a promise that resolves once all
    * operations are confirmed.
    *
-   * @see module:bitfinex-api-node.WSv2#submitOrder
+   * @see {@link WSv2#submitOrder}
    *
    * @param {(
-   *   module:bfx-api-node-models.Order[]|
-   *   module:bfx-api-node-models.Order~Data[]|
+   *   bfx-api-node-models.Order[]|
+   *   bfx-api-node-models.Order~Data[]|
    *   number[]
    * )} orders - array of order models, arrays or IDs to be cancelled
    *
@@ -2128,11 +2140,10 @@ class WSv2 extends EventEmitter {
    * In the future our mobile app will also support spawning native push
    * notifications in response to incoming ucm-notify-ui packets.
    *
-   * @see {@link module:bfx-api-node-models.Notification|Notification}
    * @throws an error if given no type or message, or the instance is not open
    *   and authenticated
    *
-   * @param {module:bfx-api-node-models.Notification~BroadcastPayload} opts -
+   * @param {bfx-api-node-models.Notification~BroadcastPayload} opts -
    *   broadcast payload
    */
   notifyUI (opts = {}) {
@@ -2200,7 +2211,7 @@ class WSv2 extends EventEmitter {
    * Registers a new callback to be called when a matching info message is
    * received.
    *
-   * @param {number} code - from {@link module:bitfinex-api-node.WSv2.info}
+   * @param {number} code - from {@link WSv2.info}
    * @param {Function} cb - callback
    */
   onInfoMessage (code, cb) {
@@ -2230,8 +2241,8 @@ class WSv2 extends EventEmitter {
    * @param {string|number} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
    * @see https://docs.bitfinex.com/v2/reference#ws-public-candle
-   * @see module:bitfinex-api-node.WSv2#subscribeCandles
-   * @see module:bitfinex-api-node.WSv2#unsubscribeCandles
+   * @see {@link WSv2#subscribeCandles}
+   * @see {@link WSv2#unsubscribeCandles}
    */
   onCandle ({ key, cbGID }, cb) {
     this._registerListener('candle', { 0: key }, Candle, cbGID, cb)
@@ -2247,8 +2258,8 @@ class WSv2 extends EventEmitter {
    * @param {string|number} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
    * @see https://docs.bitfinex.com/v2/reference#ws-public-order-books
-   * @see module:bitfinex-api-node.WSv2#subscribeOrderBook
-   * @see module:bitfinex-api-node.WSv2#unsubscribeOrderBook
+   * @see {@link WSv2#subscribeOrderBook}
+   * @see {@link WSv2#unsubscribeOrderBook}
    */
   onOrderBook ({ symbol, prec, len, cbGID }, cb) {
     this._registerListener('orderbook', {
@@ -2268,8 +2279,8 @@ class WSv2 extends EventEmitter {
    * @param {string|number} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
    * @see https://docs.bitfinex.com/v2/reference#ws-public-order-books
-   * @see module:bitfinex-api-node.WSv2#subscribeOrderBook
-   * @see module:bitfinex-api-node.WSv2#unsubscribeOrderBook
+   * @see {@link WSv2#subscribeOrderBook}
+   * @see {@link WSv2#unsubscribeOrderBook}
    */
   onOrderBookChecksum ({ symbol, prec, len, cbGID }, cb) {
     this._registerListener('ob_checksum', {
@@ -2289,8 +2300,8 @@ class WSv2 extends EventEmitter {
    * @param {string|number} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
    * @see https://docs.bitfinex.com/v2/reference#ws-public-trades
-   * @see module:bitfinex-api-node.WSv2#subscribeTrades
-   * @see module:bitfinex-api-node.WSv2#unsubscribeTrades
+   * @see {@link WSv2#subscribeTrades}
+   * @see {@link WSv2#unsubscribeTrades}
    */
   onTrades ({ symbol, pair, cbGID }, cb) {
     const id = pair || symbol || ''
@@ -2308,8 +2319,8 @@ class WSv2 extends EventEmitter {
    * @param {string|number} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
    * @see https://docs.bitfinex.com/v2/reference#ws-public-trades
-   * @see module:bitfinex-api-node.WSv2#subscribeTrades
-   * @see module:bitfinex-api-node.WSv2#unsubscribeTrades
+   * @see {@link WSv2#subscribeTrades}
+   * @see {@link WSv2#unsubscribeTrades}
    */
   onTradeEntry ({ pair, symbol, cbGID }, cb) {
     const id = pair || symbol || ''
@@ -2320,7 +2331,6 @@ class WSv2 extends EventEmitter {
    * Register a handler to be called on each personal trade `'te'` event
    *
    * @param {object} opts - options
-   * @param {string} [opts.pair] - required if no symbol specified
    * @param {string} [opts.symbol] - required if no pair specified
    * @param {string|number} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
@@ -2334,7 +2344,6 @@ class WSv2 extends EventEmitter {
    * Register a handler to be called on each personal trade `'tu'` event
    *
    * @param {object} opts - options
-   * @param {string} [opts.pair] - required if no symbol specified
    * @param {string} [opts.symbol] - required if no pair specified
    * @param {string|number} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
@@ -2352,8 +2361,8 @@ class WSv2 extends EventEmitter {
    * @param {string|number} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
    * @see https://docs.bitfinex.com/v2/reference#ws-public-ticker
-   * @see module:bitfinex-api-node.WSv2#subscribeTicker
-   * @see module:bitfinex-api-node.WSv2#unsubscribeTicker
+   * @see {@link WSv2#subscribeTicker}
+   * @see {@link WSv2#unsubscribeTicker}
    */
   onTicker ({ symbol = '', cbGID } = {}, cb) {
     const m = symbol[0] === 'f' ? FundingTicker : TradingTicker
@@ -2368,7 +2377,7 @@ class WSv2 extends EventEmitter {
    * @param {string} opts.key - key of feed to listen on
    * @param {string|number} [opts.cbGID] - callback group ID
    * @param {Function} cb - callback
-   * @see module:bitfinex-api-node.WSv2#subscribeStatus
+   * @see {@link WSv2#subscribeStatus}
    */
   onStatus ({ key = '', cbGID } = {}, cb) {
     this._registerListener('status', { 0: key }, null, cbGID, cb)
@@ -2385,7 +2394,7 @@ class WSv2 extends EventEmitter {
    * @param {string|number} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
    * @see https://docs.bitfinex.com/v2/reference#ws-auth-orders
-   * @see module:bitfinex-api-node.WSv2#auth
+   * @see {@link WSv2#auth}
    */
   onOrderSnapshot ({ symbol, id, cid, gid, cbGID }, cb) {
     this._registerListener('os', {
@@ -2407,7 +2416,7 @@ class WSv2 extends EventEmitter {
    * @param {string} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
    * @see https://docs.bitfinex.com/v2/reference#ws-auth-orders
-   * @see module:bitfinex-api-node.WSv2#submitOrder
+   * @see {@link WSv2#submitOrder}
    */
   onOrderNew ({ symbol, id, cid, gid, cbGID }, cb) {
     this._registerListener('on', {
@@ -2429,7 +2438,7 @@ class WSv2 extends EventEmitter {
    * @param {string} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
    * @see https://docs.bitfinex.com/v2/reference#ws-auth-orders
-   * @see module:bitfinex-api-node.WSv2#updateOrder
+   * @see {@link WSv2#updateOrder}
    */
   onOrderUpdate ({ symbol, id, cid, gid, cbGID }, cb) {
     this._registerListener('ou', {
@@ -2451,7 +2460,7 @@ class WSv2 extends EventEmitter {
    * @param {string} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
    * @see https://docs.bitfinex.com/v2/reference#ws-auth-orders
-   * @see module:bitfinex-api-node.WSv2#cancelOrder
+   * @see {@link WSv2#cancelOrder}
    */
   onOrderClose ({ symbol, id, cid, gid, cbGID }, cb) {
     this._registerListener('oc', {
@@ -2470,7 +2479,7 @@ class WSv2 extends EventEmitter {
    * @param {string} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
    * @see https://docs.bitfinex.com/v2/reference#ws-auth-position
-   * @see module:bitfinex-api-node.WSv2#auth
+   * @see {@link WSv2#auth}
    */
   onPositionSnapshot ({ symbol, cbGID }, cb) {
     this._registerListener('ps', { 0: symbol }, Position, cbGID, cb)
@@ -2524,7 +2533,7 @@ class WSv2 extends EventEmitter {
    * @param {string} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
    * @see https://docs.bitfinex.com/v2/reference#ws-auth-offers
-   * @see module:bitfinex-api-node.WSv2#auth
+   * @see {@link WSv2#auth}
    */
   onFundingOfferSnapshot ({ symbol, cbGID }, cb) {
     this._registerListener('fos', { 1: symbol }, FundingOffer, cbGID, cb)
@@ -2578,7 +2587,7 @@ class WSv2 extends EventEmitter {
    * @param {string} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
    * @see https://docs.bitfinex.com/v2/reference#ws-auth-credits
-   * @see module:bitfinex-api-node.WSv2#auth
+   * @see {@link WSv2#auth}
    */
   onFundingCreditSnapshot ({ symbol, cbGID }, cb) {
     this._registerListener('fcs', { 1: symbol }, FundingCredit, cbGID, cb)
@@ -2632,7 +2641,7 @@ class WSv2 extends EventEmitter {
    * @param {string} [opts.cbGID] - callback group id
    * @param {Function} cb - callback
    * @see https://docs.bitfinex.com/v2/reference#ws-auth-loans
-   * @see module:bitfinex-api-node.WSv2#auth
+   * @see {@link WSv2#auth}
    */
   onFundingLoanSnapshot ({ symbol, cbGID }, cb) {
     this._registerListener('fls', { 1: symbol }, FundingLoan, cbGID, cb)
