+++ lib/ws2_manager.js
@@ -24,8 +24,7 @@ const reconnectThrottler = new PromiseThrottle({
  * getFreeDataSocket(), or create a new WSv2 instance manually
  *
  * @class
- * @augments EventEmitter
- * @memberof module:bitfinex-api-node
+ * @augments events.EventEmitter
  * @example
  * const rest = new RESTv2()
  * const details = await rest.symbolDetails()
@@ -83,8 +82,8 @@ const reconnectThrottler = new PromiseThrottle({
  */
 class WS2Manager extends EventEmitter {
   /**
-   * @param {module:bitfinex-api-node.WSv2~Configuration} socketArgs - passed
-   *   to {@link module:bitfinex-api-node.WSv2|WSv2} constructors
+   * @param {WSv2~Configuration} socketArgs - passed to {@link WSv2}
+   *   constructor
    * @param {object} [authArgs] - cached for all internal socket auth() calls
    * @param {object} [authArgs.calc] - default 0
    * @param {object} [authArgs.dms] - default 0
@@ -146,7 +145,7 @@ class WS2Manager extends EventEmitter {
   }
 
   /**
-   * @param {module:bitfinex-api-node.WS2Manager~SocketState} s - socket state
+   * @param {WS2Manager~SocketState} s - socket state
    * @returns {number} count - # of subscribed/pending data channels
    */
   static getDataChannelCount (s) {
@@ -167,7 +166,7 @@ class WS2Manager extends EventEmitter {
 
   /**
    * @param {number} i - index into pool
-   * @returns {module:bitfinex-api-node.WS2Manager~SocketState} state
+   * @returns {WS2Manager~SocketState} state
    */
   getSocket (i) {
     return this._sockets[i]
@@ -220,7 +219,7 @@ class WS2Manager extends EventEmitter {
    * event listeners to forward via our own event emitter, and to manage pending
    * subs/unsubs.
    *
-   * @returns {module:bitfinex-api-node.WS2Manager~SocketState} state
+   * @returns {WS2Manager~SocketState} state
    */
   openSocket () {
     const { apiKey, apiSecret } = this._socketArgs
@@ -296,7 +295,7 @@ class WS2Manager extends EventEmitter {
   }
 
   /**
-   * @returns {module:bitfinex-api-node.WS2Manager~SocketState} state
+   * @returns {WS2Manager~SocketState} state
    */
   getAuthenticatedSocket () {
     return this._sockets.find(s => s.ws.isAuthenticated())
@@ -306,7 +305,7 @@ class WS2Manager extends EventEmitter {
    * Returns the first socket that has less active/pending channels than the
    * DATA_CHANNEL_LIMIT
    *
-   * @returns {module:bitfinex-api-node.WS2Manager~SocketState} state - undefined if none found
+   * @returns {WS2Manager~SocketState} state - undefined if none found
    */
   getFreeDataSocket () {
     return this._sockets.find(s => (
@@ -320,7 +319,7 @@ class WS2Manager extends EventEmitter {
    *
    * @param {string} type - i.e. 'book'
    * @param {object} filter - i.e. { symbol: 'tBTCUSD', prec: 'R0' }
-   * @returns {module:bitfinex-api-node.WS2Manager~SocketState} wsState -
+   * @returns {WS2Manager~SocketState} wsState -
    *   undefined if not found
    */
   getSocketWithDataChannel (type, filter) {
@@ -348,7 +347,7 @@ class WS2Manager extends EventEmitter {
    * NOTE: Cannot filter against pending subscriptions, due to unknown chanId
    *
    * @param {number} chanId - channel ID
-   * @returns {module:bitfinex-api-node.WS2Manager~SocketState} wsState -
+   * @returns {WS2Manager~SocketState} wsState -
    *   undefined if not found
    */
   getSocketWithChannel (chanId) {
@@ -363,7 +362,7 @@ class WS2Manager extends EventEmitter {
   /**
    * @param {string} channel - channel type
    * @param {string} identifier - unique channel identifier
-   * @returns {module:bitfinex-api-node.WS2Manager~SocketState} wsState -
+   * @returns {WS2Manager~SocketState} wsState -
    *   undefined if not found
    */
   getSocketWithSubRef (channel, identifier) {
