+++ test/index.js
@@ -5,7 +5,7 @@ const assert = require('assert')
 const BFX = require('../index')
 const { RESTv1, RESTv2 } = require('bfx-api-node-rest')
 const WSv1 = require('bfx-api-node-ws1')
-const WSv2 = require('../lib/transports/ws2')
+const WSv2 = require('../lib/ws2')
 
 describe('BFX', () => {
   it('should be loaded', () => {
